splitting the components
